fix: Refactored git url commit to function
